docker build: accept -f - to read Dockerfile from stdin
Implements proposal #27393 that describes the meaning of '-f -' in a
`docker build` command. For all context types except reading a context from
stdin, '-f -' will replace any file called 'Dockerfile' in the context's
tar stream with the contents of stdin. If the context itself is read from
stdin, '-f -' continues to read the Dockerfile called '-' from this context.

Fixes #27393.

Signed-off-by: David Sheets <sheets@alum.mit.edu>