volume: moved source path existence validation to mount failure
This simplifies mount spec validation by removing the file system state
input side effect of checking whether the source path exists. Instead,
this error condition is naturally handled during bind mounting. To
achieve this, I introduce a new flag on volume.MountPoint,
CreateSourceIfMissing, that indicates that the source path should be
created if it is missing. We use this flag to decide whether to
idtools.MkdirAllAndChownNew rather than always executing that function
and relying on the early spec validation to prevent --mount requests
from creating directory hierarchies. I also simplified
volume.MountPoint.Setup to remove the optional checking function as this
check can now be performed accurately immediately after
volume.MountPoint.Realize for Unix-like systems.

This change was necessary in order to support mount point plugins that
implement full file system interposition as the requested source should
not be checked early in that case and, instead, the effective source
should be used for mounting (or fail).

Signed-off-by: David Sheets <dsheets@docker.com>